move isExpression and create template regex to lib/expressions
